+++ sourcefiles/cf-bootstrap-jquery-login/js/layout/mainlayout.js
@@ -1,0 +1,6 @@
+/**
+ * Function to handle logout action.
+ */
+function onLogout() {
+    window.location.href = "index.cfm?page=logout";
+}