+++ CHANGELOG.md
@@ -6,16 +6,26 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ---
 
+## [1.3.2] - 2025-10-02
+### Added
+- Added Adobe ColdFusion Project Template login with Bootstrap/jQuery inspired by Blazor Web App.
+  Generates a Adobe ColdFusion project structure template with:
+  - Bootstrap 5.3.6, Bootstrap Icons 1.11.3 and jQuery 3.7.1 preconfigured.
+  - Sample pages with login and simple session handling.
+  - Sample pages inspired by a Blazor Web App.
+
+---
+
 ## [1.2.2] - 2025-06-22
 ### Added
 - Minor changes
-- Add badges to extension.
+- Added badges to extension.
 
 ---
 
 ## [1.1.2] - 2025-06-22
 ### Added
-- Add CHANGELOG file.
+- Added CHANGELOG file.
 
 ---
 
@@ -29,7 +39,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Initial release
 ## [1.0.1] - 2025-06-21
 ### Added
-- Initial release of the Adobe ColdFusion Project Template with Bootstrap and jQuery inspired by Blazor Web App.
+- Initial release of the Adobe ColdFusion Project Template with Bootstrap/jQuery inspired by Blazor Web App.
 - Generates a Adobe ColdFusion project structure template with:
   - Bootstrap 5.3.6 and jQuery 3.7.1 preconfigured.
   - Sample pages inspired by a Blazor Web App.